+++ src/components/InstantTrade.jsx
@@ -7,17 +7,31 @@ import React, {
 import { useTranslation } from "react-i18next";
 import { i18n as i18nInstance, locale } from "@/lib/i18n.js";
 import { useStore } from "@nanostores/react";
+import { ReloadIcon } from "@radix-ui/react-icons";
 
-import InstantTradeMarket from "./InstantTradeMarket";
+import { humanReadableFloat, trimPrice, isInvertedMarket } from "@/lib/common";
 
-import { humanReadableFloat } from "@/lib/common";
 import { useInitCache } from "@/nanoeffects/Init.ts";
 import { createUserBalancesStore } from "@/nanoeffects/UserBalances.ts";
+import { createAssetFromSymbolStore } from "@/nanoeffects/Assets.ts";
+import { createMarketOrderStore } from "@/nanoeffects/MarketOrderBook.ts";
 
 import { $currentUser } from "@/stores/users.ts";
 import { $currentNode } from "@/stores/node.ts";
 
-import { createAssetFromSymbolStore } from "@/nanoeffects/Assets.ts";
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardFooter,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card";
+import { Button } from "@/components/ui/button";
+
+import HorizontalTradeCard from "./Market/HorizontalTradeCard.jsx";
+import MarketOrderCard from "./Market/MarketOrderCard.jsx";
+import AssetDropDown from "./Market/AssetDropDownCard.jsx";
 
 export default function InstantTrade(properties) {
   const { t, i18n } = useTranslation(locale.get(), { i18n: i18nInstance });
@@ -112,17 +126,11 @@ export default function InstantTrade(properties) {
     () => marketSearch.map((asset) => asset.s),
     [marketSearch]
   );
+
   const searchIds = useMemo(
     () => marketSearch.map((asset) => asset.id),
     [marketSearch]
   );
-
-  const handleAssetAChange = (newAssetA) => {
-    setAssetA(newAssetA);
-  };
-  const handleAssetBChange = (newAssetA) => {
-    setAssetB(newAssetA);
-  };
 
   const [assetA, setAssetA] = useState(!window.location.search ? "BTS" : null);
   const [assetB, setAssetB] = useState(!window.location.search ? "CNY" : null);
@@ -268,6 +276,186 @@ export default function InstantTrade(properties) {
     }
   }, [assets, assetB, usr]);
 
+  const [usrBalances, setUsrBalances] = useState(null);
+  useEffect(() => {
+    async function fetchUsrBalances() {
+      if (usr && usr.id) {
+        const userBalancesStore = createUserBalancesStore([
+          usr.chain,
+          usr.id,
+          currentNode ? currentNode.url : null,
+        ]);
+
+        userBalancesStore.subscribe(({ data, error, loading }) => {
+          if (data && !error && !loading) {
+            const filteredData = data.filter((balance) =>
+              assets.find((x) => x.id === balance.asset_id)
+            );
+            setUsrBalances(filteredData);
+          }
+        });
+      }
+    }
+
+    fetchUsrBalances();
+  }, [usr]);
+
+  const invertedMarket = useMemo(() => {
+    return isInvertedMarket(assetAData.id, assetBData.id);
+  }, [assetAData, assetBData]);
+
+  const [buyOrders, setBuyOrders] = useState(null);
+  const [sellOrders, setSellOrders] = useState(null);
+  useEffect(() => {
+    async function fetchMarketOrders() {
+      const marketOrdersStore = createMarketOrderStore([
+        usr.chain,
+        assetA,
+        assetB,
+        50,
+      ]);
+
+      marketOrdersStore.subscribe(({ data, error, loading }) => {
+        if (data && !error && !loading) {
+          setBuyOrders(data.bids);
+          setSellOrders(data.asks);
+        } else {
+          setBuyOrders(null);
+          setSellOrders(null);
+        }
+      });
+    }
+
+    if (usr && assetA && assetB) {
+      fetchMarketOrders();
+    }
+  }, [usr, assetA, assetB]);
+
+  const [clicked, setClicked] = useState(false);
+
+  const [amount, setAmount] = useState(0.0);
+  const [price, setPrice] = useState(0.0);
+  const [total, setTotal] = useState(0);
+
+  const marketFees = useMemo(() => {
+    let calculatedMarketFee = 0.0;
+
+    if (amount && price && total) {
+      if (
+        orderType === "buy" &&
+        assetAData &&
+        assetAData.market_fee_percent &&
+        assetAData.market_fee_percent > 0
+      ) {
+        calculatedMarketFee =
+          parseFloat(amount) * (assetAData.market_fee_percent / 100);
+        return calculatedMarketFee.toFixed(assetAData.precision);
+      }
+
+      if (
+        orderType === "sell" &&
+        assetBData &&
+        assetBData.market_fee_percent &&
+        assetBData.market_fee_percent > 0
+      ) {
+        calculatedMarketFee =
+          parseFloat(total) * (assetBData.market_fee_percent / 100);
+        return calculatedMarketFee.toFixed(assetBData.precision);
+      }
+    }
+
+    return calculatedMarketFee;
+  }, [amount, price, total, orderType, assetAData, assetBData]);
+
+  const [expiryType, setExpiryType] = useState("fkill");
+  const [expiry, setExpiry] = useState(() => {
+    const now = new Date();
+    const oneHour = 60 * 60 * 1000;
+    return new Date(now.getTime() + oneHour);
+  });
+
+  const [date, setDate] = useState(
+    new Date(Date.now() + 1 * 24 * 60 * 60 * 1000)
+  );
+
+  const form = useForm({
+    defaultValues: {
+      priceAmount: 0.0,
+      sellAmount: 0.0,
+      sellTotal: 0,
+      expiry: "fkill",
+      fee: 0,
+      marketFees: 0,
+    },
+  });
+
+  const [showDialog, setShowDialog] = useState(false);
+
+  function getReadableBalance(assetData, balances) {
+    const id = assetData.id;
+    const foundBalance = balances.find((x) => x.asset_id === id);
+    return foundBalance
+      ? humanReadableFloat(
+          foundBalance.amount,
+          assetData.precision
+        ).toLocaleString(undefined, {
+          minimumFractionDigits: assetData.precision,
+        })
+      : 0;
+  }
+
+  const assetABalance = useMemo(() => {
+    return assetAData && usrBalances
+      ? getReadableBalance(assetAData, usrBalances)
+      : 0;
+  }, [assetAData, usrBalances]);
+
+  const assetBBalance = useMemo(() => {
+    return assetBData && usrBalances
+      ? getReadableBalance(assetBData, usrBalances)
+      : 0;
+  }, [assetBData, usrBalances]);
+
+  const [inputChars, setInputChars] = useState(0);
+  useEffect(() => {
+    if (inputChars > 0) {
+      let finalUrlParams =
+        `?market=${thisAssetA}_${thisAssetB}` + `&amount=${amount}`;
+
+      window.history.replaceState({}, "", finalUrlParams);
+    }
+  }, [amount, thisAssetA, thisAssetB]);
+
+  const trxJSON = useMemo(() => {
+    // TODO: process limit orders which match multiple open market orders
+    return [
+      {
+        seller: usr.id,
+        amount_to_sell: {
+          amount: blockchainFloat(total, assetBData.precision).toFixed(0),
+          asset_id: marketSearch.find((asset) => asset.s === thisAssetB).id,
+        },
+        min_to_receive: {
+          amount: blockchainFloat(amount, assetAData.precision).toFixed(0),
+          asset_id: marketSearch.find((asset) => asset.s === thisAssetA).id,
+        },
+        expiration: date,
+        fill_or_kill: true,
+        extensions: {},
+      },
+    ];
+  }, [
+    usr,
+    total,
+    assetBData,
+    thisAssetB,
+    amount,
+    assetAData,
+    thisAssetA,
+    date,
+    marketSearch,
+  ]);
+
   if (
     !usr ||
     !usr.chain ||
@@ -287,30 +475,405 @@ export default function InstantTrade(properties) {
     );
   }
 
+  const Row = ({ index, style }) => {
+    const order = marketOrders[index];
+
+    const price = parseFloat(order.price).toFixed(assetBData.precision);
+    const base = parseFloat(order.base);
+    const quote = parseFloat(order.quote);
+
+    const totalBase = marketOrders
+      .slice(0, index + 1)
+      .map((x) => parseFloat(x.base))
+      .reduce((acc, curr) => acc + curr, 0)
+      .toFixed(assetBData.precision);
+
+    const totalQuote = marketOrders
+      .slice(0, index + 1)
+      .map((x) => parseFloat(x.quote))
+      .reduce((acc, curr) => acc + curr, 0)
+      .toFixed(assetAData.precision);
+
+    return (
+      <div style={style}>
+        <div className="grid grid-cols-4 text-sm">
+          <div className="col-span-1 border-l-2 border-r-2 pl-3 font-mono text-right tabular-nums">
+            {price}
+          </div>
+          <div className="col-span-1 border-r-2 pl-3 font-mono text-right tabular-nums">
+            {cardType === "buy"
+              ? base.toFixed(assetBData.precision)
+              : quote.toFixed(assetAData.precision)}
+          </div>
+          <div className="col-span-1 border-r-2 pl-3 font-mono text-right tabular-nums">
+            {cardType === "buy"
+              ? quote.toFixed(assetAData.precision)
+              : base.toFixed(assetBData.precision)}
+          </div>
+          <div className="col-span-1 pl-3 font-mono text-right tabular-nums">
+            {totalBase}
+          </div>
+        </div>
+      </div>
+    );
+  };
+
   return (
     <div className="container mx-auto mt-5 mb-5">
-      <InstantTradeMarket
-        usr={usr}
-        assetA={assetA}
-        assetB={assetB}
-        assetAData={assetAData}
-        assetADetails={assetADetails}
-        assetABitassetData={aBitassetData}
-        assetBData={assetBData}
-        assetBDetails={assetBDetails}
-        assetBBitassetData={bBitassetData}
-        limitOrderFee={limitOrderFee}
-        setAssetA={handleAssetAChange}
-        setAssetB={handleAssetBChange}
-        key={`Market_${assetA}_${assetB}`}
-        _assetsBTS={_assetsBTS}
-        _assetsTEST={_assetsTEST}
-        _marketSearchBTS={_marketSearchBTS}
-        _marketSearchTEST={_marketSearchTEST}
-        _poolsBTS={_poolsBTS}
-        _poolsTEST={_poolsTEST}
-        balances={balances}
-      />
+      <div className="grid grid-cols-1 md:grid-cols-2 gap-5">
+        <div className="col-span-1">
+          <div className="flex-grow mb-2">
+            <Card>
+              <CardHeader className="pt-2 pb-2">
+                <CardTitle className="text-lg">
+                  {usr.chain === "bitshares"
+                    ? "Bitshares "
+                    : "Bitshares (Testnet) "}
+                  Instant Trade
+                </CardTitle>
+              </CardHeader>
+              <CardContent className="pb-3">
+                <div className="grid grid-cols-3 gap-1">
+                  <AssetDropDown
+                    assetSymbol={assetA}
+                    assetData={assetAData}
+                    storeCallback={setAssetA}
+                    otherAsset={assetB}
+                    marketSearch={marketSearch}
+                    type={"quote"}
+                    size="small"
+                    chain={usr.chain}
+                    balances={balances}
+                  />
+
+                  <a
+                    style={{ lineHeight: 1 }}
+                    href={`/dex/index.html?market=${assetB}_${assetA}`}
+                    onClick={() => setClicked(true)}
+                  >
+                    <Button variant="outline" className="w-full h-7">
+                      {clicked ? (
+                        <ReloadIcon className="animate-spin" />
+                      ) : (
+                        <ReloadIcon />
+                      )}
+                    </Button>
+                  </a>
+
+                  <AssetDropDown
+                    assetSymbol={assetB}
+                    assetData={assetBData}
+                    storeCallback={setAssetB}
+                    otherAsset={assetA}
+                    marketSearch={marketSearch}
+                    type={"base"}
+                    size="small"
+                    chain={usr.chain}
+                    balances={balances}
+                  />
+                </div>
+              </CardContent>
+            </Card>
+          </div>
+
+          <Card>
+            <CardHeader className="pb-2">
+              <CardTitle>
+                {t("LimitOrderCard:buyingWith", {
+                  assetA: thisAssetA,
+                  assetB: thisAssetB,
+                })}
+              </CardTitle>
+              <CardDescription>
+                {t("LimitOrderCard:createLimitOrder")}
+              </CardDescription>
+            </CardHeader>
+            <CardContent>
+              {thisAssetA &&
+              thisAssetB &&
+              marketSearch &&
+              assetAData &&
+              assetBData ? (
+                <form onSubmit={form.handleSubmit(() => setShowDialog(true))}>
+                  <FieldGroup>
+                    <Controller
+                      name="sellAmount"
+                      control={form.control}
+                      render={({ field, fieldState }) => (
+                        <Field
+                          invalid={fieldState.invalid}
+                          className="mt-4 text-xs"
+                        >
+                          <FieldLabel>
+                            {t("LimitOrderCard:sellAmount.label")}
+                          </FieldLabel>
+                          <FieldDescription>
+                            {t("LimitOrderCard:sellAmount.buyDescription", {
+                              asset: thisAssetA,
+                            })}
+                          </FieldDescription>
+                          <span className="grid grid-cols-12">
+                            <span className="col-span-9">
+                              <Input
+                                {...field}
+                                label={`Amount`}
+                                placeholder={amount}
+                                disabled
+                                readOnly
+                              />
+                            </span>
+                            <span className="col-span-3 ml-3 text-center">
+                              <Popover>
+                                <PopoverTrigger>
+                                  <span className="inline-block border border-gray-300 rounded pl-4 pb-1 pr-4 text-lg">
+                                    <Label>
+                                      {t("LimitOrderCard:editLabel")}
+                                    </Label>
+                                  </span>
+                                </PopoverTrigger>
+                                <PopoverContent>
+                                  <Label>
+                                    {t(
+                                      "LimitOrderCard:sellAmount.provideNewLabel"
+                                    )}
+                                  </Label>{" "}
+                                  <Input
+                                    placeholder={amount}
+                                    className="mb-2 mt-1"
+                                    onChange={(event) => {
+                                      const input = event.target.value;
+                                      const regex = /^[0-9,]*\.?[0-9]*$/;
+                                      if (
+                                        input &&
+                                        input.length &&
+                                        regex.test(input)
+                                      ) {
+                                        const parsedInput = parseFloat(
+                                          input.replaceAll(",", "")
+                                        );
+                                        if (parsedInput) {
+                                          setAmount(
+                                            parsedInput.toFixed(
+                                              assetAData.precision
+                                            )
+                                          );
+                                          if (price) {
+                                            setTotal(
+                                              (parsedInput * price).toFixed(
+                                                assetBData.precision
+                                              )
+                                            );
+                                          }
+                                          setInputChars(inputChars + 1);
+                                        }
+                                      }
+                                    }}
+                                  />
+                                </PopoverContent>
+                              </Popover>
+                            </span>
+                          </span>
+                        </Field>
+                      )}
+                    />
+
+                    <Separator className="mt-3" />
+
+                    <Controller
+                      name="fee"
+                      control={form.control}
+                      render={({ field, fieldState }) => (
+                        <Field invalid={fieldState.invalid} disabled>
+                          <FieldLabel>
+                            {t("LimitOrderCard:fee.label")}
+                          </FieldLabel>
+                          <FieldDescription>
+                            {t("LimitOrderCard:fee.description")}
+                          </FieldDescription>
+                          <Input
+                            {...field}
+                            disabled
+                            label={t("LimitOrderCard:fee.label")}
+                            value={`${fee} BTS`}
+                            placeholder={1}
+                          />
+                          {expiryType === "fkill" || usr.id === usr.referrer ? (
+                            <FieldError>
+                              {expiryType === "fkill"
+                                ? t("LimitOrderCard:fee.unfilledRebate", {
+                                    fee,
+                                  })
+                                : null}
+                              <br />
+                              {usr.id === usr.referrer
+                                ? t("LimitOrderCard:fee.ltmRebate", {
+                                    rebate: 0.8 * fee,
+                                  })
+                                : null}
+                            </FieldError>
+                          ) : null}
+                        </Field>
+                      )}
+                    />
+                    {orderType === "buy" &&
+                    assetAData &&
+                    assetAData.market_fee_percent &&
+                    assetAData.market_fee_percent > 0 ? (
+                      <Controller
+                        name="marketFees"
+                        control={form.control}
+                        render={({ field, fieldState }) => (
+                          <Field invalid={fieldState.invalid} disabled>
+                            <FieldLabel>
+                              {t("LimitOrderCard:marketFees.label")}
+                            </FieldLabel>
+                            <Input
+                              {...field}
+                              disabled
+                              value={`${marketFees} ${assetAData.symbol}`}
+                              placeholder={`${marketFees} ${assetAData.symbol}`}
+                            />
+                            <FieldDescription>
+                              {t("LimitOrderCard:marketFees.description")}
+                            </FieldDescription>
+                          </Field>
+                        )}
+                      />
+                    ) : null}
+                    {!amount || !price || !expiry ? (
+                      <Button
+                        className="mt-7 mb-1"
+                        variant="outline"
+                        disabled
+                        type="submit"
+                      >
+                        {t("LimitOrderCard:submit")}
+                      </Button>
+                    ) : (
+                      <Button
+                        className="mt-7 mb-1"
+                        variant="outline"
+                        type="submit"
+                      >
+                        {t("LimitOrderCard:submit")}
+                      </Button>
+                    )}
+                  </FieldGroup>
+                </form>
+              ) : null}
+            </CardContent>
+
+            <Card>
+              <CardHeader>
+                <CardTitle>
+                  {cardType === "buy"
+                    ? t("MarketOrderCard:openBuyLimitOrdersTitle")
+                    : t("MarketOrderCard:openSellLimitOrdersTitle")}
+                </CardTitle>
+                <CardDescription>
+                  {t(
+                    cardType === "buy"
+                      ? "MarketOrderCard:buyLimitOrdersDescription"
+                      : "MarketOrderCard:sellLimitOrdersDescription",
+                    {
+                      assetA: assetA,
+                      assetB: assetB,
+                    }
+                  )}
+                </CardDescription>
+              </CardHeader>
+              <CardContent>
+                {marketOrders && marketOrders.length ? (
+                  <>
+                    <div className="grid grid-cols-4">
+                      <div className="col-span-1 pl-3 text-right pr-2">
+                        Price
+                      </div>
+                      <div className="col-span-1 pl-3 text-md text-right pr-2">
+                        {cardType === "sell" && assetA && assetA.length < 12
+                          ? assetA
+                          : null}
+                        {cardType === "sell" &&
+                        assetA &&
+                        assetA.length >= 12 &&
+                        assetAData
+                          ? assetAData.id
+                          : null}
+                        {cardType === "buy" && assetB && assetB.length < 12
+                          ? assetB
+                          : null}
+                        {cardType === "buy" &&
+                        assetB &&
+                        assetB.length >= 12 &&
+                        assetBData
+                          ? assetBData.id
+                          : null}
+                      </div>
+                      <div className="col-span-1 pl-3 text-right pr-2">
+                        {cardType === "sell" && assetB && assetB.length < 12
+                          ? assetB
+                          : null}
+                        {cardType === "sell" &&
+                        assetB &&
+                        assetB.length >= 12 &&
+                        assetBData
+                          ? assetBData.id
+                          : null}
+                        {cardType === "buy" && assetA && assetA.length < 12
+                          ? assetA
+                          : null}
+                        {cardType === "buy" &&
+                        assetA &&
+                        assetA.length >= 12 &&
+                        assetAData
+                          ? assetAData.id
+                          : null}
+                      </div>
+                      <div className="col-span-1 pl-3 text-right pr-2">
+                        {assetB && assetB.length < 7
+                          ? `Total (${assetB})`
+                          : null}
+                        {assetB && assetB.length >= 7 && assetBData
+                          ? `Total (${assetBData.id})`
+                          : null}
+                      </div>
+                    </div>
+                    <div className="h-[300px] overflow-hidden">
+                      <List
+                        height={300}
+                        rowComponent={Row}
+                        rowCount={marketOrders.length}
+                        rowHeight={20}
+                        rowProps={{}}
+                      />
+                    </div>
+                  </>
+                ) : (
+                  t("MarketOrderCard:noOpenOrders")
+                )}
+              </CardContent>
+            </Card>
+
+            {showDialog ? (
+              <DeepLinkDialog
+                operationNames={["limit_order_create"]}
+                username={usr.username}
+                usrChain={usr.chain}
+                userID={usr.id}
+                dismissCallback={setShowDialog}
+                key={`Buying${amount}${thisAssetA}for${total}${thisAssetB}`}
+                headerText={t("LimitOrderCard:headerText.buying", {
+                  amount,
+                  thisAssetA,
+                  total,
+                  thisAssetB,
+                })}
+                trxJSON={trxJSON}
+              />
+            ) : null}
+          </Card>
+        </div>
+      </div>
     </div>
   );
 }
